Update Ex338 module for Phoenix 1.3 structure
* Add Ex338.Application consistent with Phx 1.3 folder structure
* Remove coveralls config since it is no longer used
* Add runtime_tools like the new Phx 1.3 projects
* See #309
